Code Update to 1.16

+++ src/main/java/com/bioast/addworms/utils/helpers/ParticleHelper.java
@@ -27,7 +27,8 @@ public final class ParticleHelper {
                 double d0 = random.nextGaussian() * 0.02D;
                 double d1 = random.nextGaussian() * 0.02D;
                 double d2 = random.nextGaussian() * 0.02D;
-                worldIn.addParticle((IParticleData) particleType, (double)((float)posIn.getX() + random.nextFloat()), (double)posIn.getY() + (double)random.nextFloat() * blockstate.getShape(worldIn, posIn).getEnd(Direction.Axis.Y), (double)((float)posIn.getZ() + random.nextFloat()), d0, d1, d2);
+                worldIn.addParticle((IParticleData) particleType, (float) posIn.getX() + random.nextFloat(),
+                        (double) posIn.getY() + (double) random.nextFloat() * blockstate.getShape(worldIn, posIn).getEnd(Direction.Axis.Y), (float) posIn.getZ() + random.nextFloat(), d0, d1, d2);
             }
 
         }
